refactor: 안내문구 모달 공통 컴포넌트로 변경
