+++ src/pages/signUp/components/SignupInputForm.tsx
@@ -9,7 +9,7 @@ import {
   LoadingProps,
 } from '../signupType';
 import VSignupInputForm from './VSignupInputForm';
-import SignupGuideModal from './SignupGuideModal';
+import SignupGuideModal from '../../../commons/modals/TextGuideModal';
 
 const SignupInputForm = () => {
   const [shelterInfo, setShelterInfo] = useRecoilState(shelterSignupState);
